% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaged_bootstrap_graph.R
\name{averaged_bootstrap_graph}
\alias{averaged_bootstrap_graph}
\title{Create Partial Ancestral Graph Based on Bootrstrap Analysis}
\usage{
averaged_bootstrap_graph(
  Arrowheads_results,
  Tails_results,
  Circles_results,
  Edges_results,
  sig = FALSE,
  edge_threshold = 0,
  mark_threshold = 0
)
}
\arguments{
\item{Arrowheads_results}{Dataframe holding the proportion of time each arrowhead has been included created by bootstrap_results(_, "Arrowheads")}

\item{Tails_results}{Dataframe holding the proportion of time each tail has been included created by bootstrap_results(_, "Tails")}

\item{Circles_results}{Dataframe holding the proportion of time each circle has been included created by bootstrap_results(_, "Circles")}

\item{Edges_results}{Dataframe holding the proportion of time each edge has been included created by bootstrap_results(_, "Edges")}

\item{sig}{Logical if true only features are included that are identified significantly more often than 0 times.}

\item{edge_threshold}{Value between 0 and 1. Only edges are included with a proportion significantly greater than the threshold.}

\item{mark_threshold}{Value between 0 and 1. Only marks are included with a proportion significantly greater than the threshold.}
}
\description{
Plots a pag based on the bootstrap analysis. First all edges that are identified
in the bootstrap samples significantly more often than the threshold are included.
Afterwards all marks for the ends of the edges that are identified
in the bootstrap samples significantly more often than the threshold are included.
The remaining edge ends that do not have a mark are assignned a circle to
illustrate the observed uncertainity in these marks.
}
\examples{

# averaged_bootstrap_graph(
#   Arrowheads_results = Arrowheads_results,
#   Tails_results = Tails_results,
#   Circles_results = Circles_results,
#   Edges_results = Edges_results,
#   edge_threshold = .8,
#   mark_threshold = .8
# )

}
