% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_results.R
\name{bootstrap_results}
\alias{bootstrap_results}
\title{Extract Results from FCI Bootstrap Object}
\usage{
bootstrap_results(bootstrap_samples, feature, CI = TRUE, alpha = 0.05)
}
\arguments{
\item{bootstrap_samples}{bootstrap_samples object created by fci_bootstrap.}

\item{feature}{Feature for which results should be extracted. Options are: "Edges", "Arrowheads", "Tails", and "Circles".}

\item{CI}{Logical whether CI for proportions should be calculated.}

\item{alpha}{Alpha level that should be used for the calculated confidence intervals.}
}
\value{
A data frame holding the proportion of times each featre has been identified and a corresponding confidence interval.
}
\description{
Extracts proportion of time a graph feature has been identified by a causal inference algorithm
within the bootstrap samples created by fci_bootsrap.
}
\examples{
# Edges_results <- bootstrap_results(
#   bootstrap_samples = bootstrap_samples,
#   feature = "Edges"
# )
# Arrowheads_results <- bootstrap_results(
#   bootstrap_samples = bootstrap_samples,
#   feature = "Arrowheads"
# )
# Tails_results <- bootstrap_results(
#   bootstrap_samples = bootstrap_samples,
#   feature = "Tails"
# )
# Circles_results <- bootstrap_results(
#   bootstrap_samples = bootstrap_samples,
#   feature = "Circles"
# )

}
