% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci_precision.R
\name{check_input}
\alias{check_input}
\title{Helper function to check the input of fci_precision and return appropriate error messages.
Does nothing else.}
\usage{
check_input(
  true_graph,
  labels,
  beta,
  target,
  cycle_nodes,
  alpha,
  nmax,
  nmin,
  steps,
  n_iter,
  plot_line,
  layout,
  verbose
)
}
\description{
Helper function to check the input of fci_precision and return appropriate error messages.
Does nothing else.
}
