% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci_precision.R
\name{compare_graphs}
\alias{compare_graphs}
\title{Helper function to compare a graph generated by FCI with an expected optimal graph,
to be used within fci_precision}
\usage{
compare_graphs(probe, target, cycle_nodes, labels)
}
\arguments{
\item{probe}{The output of fci.}

\item{target}{Target partial ancestral graph that optimally should be returned by FCI in the form of a pcalg pag amat matrix.}

\item{cycle_nodes}{Character vector of nodes that are part of a cycle.}

\item{lables}{Character vector of variable names or node labels.}
}
\value{
Returns a list with feature sensitivities and orientation precision.
}
\description{
Helper function to compare a graph generated by FCI with an expected optimal graph,
to be used within fci_precision
}
