% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci_bootstrap.R
\name{fci_bootstrap}
\alias{fci_bootstrap}
\title{Bootstrap Sampling for the Fast Causal Inference Algorithm}
\usage{
fci_bootstrap(
  data,
  labels,
  amount = 100,
  sample_proportion = 1,
  replace = TRUE,
  alpha = 0.01
)
}
\arguments{
\item{data}{Data frame of the variables to be analysed.}

\item{labels}{Character vector of variable names or node labels.}

\item{amount}{Amount of bootstrap samples to be drawn.}

\item{sample_proportion}{Relative sample size of the bootstrap samples to the observed data.}

\item{replace}{Logical indicating whether resampling should be done with or without replacement. Needs to be true for sample_proportion > 1.}

\item{alpha}{Alpha value for significance tests within FCI.}
}
\value{
Returns data frame holding all bootstrap samples
}
\description{
Bootstrap analysis of FCI results for a given data set.
Can be used to quantify the uncertainty that is present for any graph feature within the data.
}
\examples{

# bootstrap_n <- 200
# sample_proportion <- 1
# threshold <- 1
# bootstrap_samples <- fci_bootstrap(
#   data = data, labels = colnames(data), amount = bootstrap_n,
#   sample_proportion = sample_proportion
# )

}
