% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci_precision.R
\name{fci_precision}
\alias{fci_precision}
\title{Sensitivity Simulations for Directed Linear and Gaussian Causal Systems}
\usage{
fci_precision(
  true_graph,
  labels,
  beta,
  target,
  cycle_nodes = NULL,
  alpha = 0.05,
  nmax = 800,
  nmin = 50,
  steps = 20,
  n_iter = 30,
  plot_line = TRUE,
  layout = "complete_m",
  verbose = FALSE
)
}
\arguments{
\item{true_graph}{The causal structure of the system that is simulated from in form of a directed adjacency matrix.}

\item{labels}{Character vector of variable names or node labels.}

\item{beta}{Mean value for a normal distribution from which edge slopes are drawn.}

\item{target}{Target partial ancestral graph that optimally should be returned by FCI in the form of a pcalg pag amat matrix.}

\item{cycle_nodes}{Character vector of nodes that are part of a cycle.}

\item{alpha}{Alpha value for significance tests within FCI.}

\item{nmax}{Maximum sample size for the simulation.}

\item{nmin}{Minimum sample size for the simulation.}

\item{steps}{Steps of sample sizes to be simulated between the minimum and maximum.}

\item{n_iter}{Amount of samples drawn at each sample size.}

\item{plot_line}{Logical whether loess regression line should be added to result plot.}

\item{layout}{Layout for the plots illustrating the results of the simulation. Options are: "complete_m", "cycle_m", "dag_m", "complete_s", "cycle_s", "dag_s", "edges_s".}

\item{verbose}{Verbose argument for the FCI function}
}
\value{
A list with a data frame holding the complete simulation and a data frame holding all results at each sample size. Additionally, plots the simulation results.
}
\description{
Simulate data from a potentially cyclical, directed, linear and Gaussian system at
different sample sizes. Apply the FCI algorithm to each sample and analyse
how often each feature of the graph has been identified correctly at each sample size.
}
\examples{
# # Nr of variables
# p <- 5
#
#
# # True SCM
# true_graph <- matrix(c(
#   0, 1, 0, 0, 0,
#   0, 0, 1, 0, 0,
#   0, 0, 0, 1, 0,
#   0, 0, 0, 0, 1,
#   0, 1, 0, 0, 0
# ), ncol = p, byrow = TRUE)
#
# # Labels
# labels <- c("A", "B", "C", "D", "E")
#
# # Define Betas to Go Through
# beta_graph <- c(.5, 1.1, 2, 3, 5)
#
# # Define Sample Range
# nmin <- 20
# nmax <- 800
# steps <- 20
# n_iter <- 80
#
# # Define Cycle Nodes
# cycle_nodes <- c("B", "C", "D", "E")
#
# # Define Ploting Parameters
# plot_line <- TRUE
# layout <- "cycle_m"
#
# target <- matrix(c(
#   0, 2, 0, 0, 2,
#   1, 0, 1, 0, 1,
#   0, 1, 0, 1, 0,
#   0, 0, 1, 0, 1,
#   1, 1, 0, 1, 0
# ), ncol = p, byrow = TRUE)
#
# # Run Simulation
# fci_precision(
#   true_graph = true_graph, labels = labels,
#   beta = beta_graph, target = target, nmin = nmin, nmax = nmax,
#   steps = steps, n_iter = n_iter, cycle_nodes = cycle_nodes,
#   plot_line = plot_line, layout = layout
# )

}
