% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci_precision.R
\name{format_results}
\alias{format_results}
\title{Helper function to format results within fci_precision.}
\usage{
format_results(simulation_grid)
}
\arguments{
\item{simulation_grid}{Data frame holding the simulation results created within fci_precision.}
}
\value{
A list of two data frames holding the simulation results in long format for plotting.
}
\description{
Helper function to format results within fci_precision.
}
