% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_averaged_bootstrap_graph.R
\name{make_averaged_bootstrap_graph}
\alias{make_averaged_bootstrap_graph}
\title{Plot averaged partial ancestral graph based on bootstrap analysis}
\usage{
make_averaged_bootstrap_graph(
  Arrowheads_results,
  Tails_results,
  Circles_results,
  Edges_results,
  sig = FALSE,
  edge_threshold = 0,
  mark_threshold = 0
)
}
\arguments{
\item{Arrowheads_results}{Dataframe holding the results of the bootstrap analysis for the arrowheads.}

\item{Tails_results}{Dataframe holding the results of the bootstrap analysis for the tails.}

\item{Circles_results}{Dataframe holding the results of the bootstrap analysis for the circles.}

\item{Edges_results}{Dataframe holding the results of the bootstrap analysis for the edges.}

\item{sig}{Boolean to indicate whether only significant edges should be plotted.}

\item{edge_threshold}{If sig=False removes all edges that occur in a smaller proportion of bootstrap samples than this.}

\item{mark_threshold}{If sig=False removes all marks that occur in a smaller proportion of bootstrap samples than this.}
}
\description{
Function to create a averaged partial ancestral graph based on the bootstrap analysis.
For this first all edges are included that occur in a larger proportion of
bootstrap samples than a preset criterion. Then all marks are applied to the
ends of these edges that occur in a larger proportion of
bootstrap samples than a preset criterion. Lastly, circles are assigned to the remaining edge
to indicate missing evidence.
}
\examples{

# make_averaged_bootstrap_graph(
#   Arrowheads_results = Arrowheads_results_without,
#   Tails_results = Tails_results_without,
#   Circles_results = Circles_results_without,
#   Edges_results = Edges_results_without,
#   edge_threshold = .8,
#   mark_threshold = .8
# )

}
