% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci_precision.R
\name{make_samples}
\alias{make_samples}
\title{Helper function to sample from a potentially cyclic, linear, Gassian graph,
to be used within fci_precision.}
\usage{
make_samples(Beta, N, true_graph)
}
\arguments{
\item{Beta}{Mean value for a normal distribution from which edge slopes are drawn.}

\item{N}{Sample size}

\item{true_graph}{The causal structure of the system that is simulated from in form of a directed adjacency matrix.}
}
\value{
A data frame with a sample based on the true graph.
}
\description{
Helper function to sample from a potentially cyclic, linear, Gassian graph,
to be used within fci_precision.
}
