% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_results.R
\name{plot_bootstrap_results}
\alias{plot_bootstrap_results}
\title{Plot results of the bootstrap graph feature analysis}
\usage{
plot_bootstrap_results(bootstrap_results_df, feature, criterion = 0.8)
}
\arguments{
\item{bootstrap_results_df}{Dataframe holding the results of the bootstrap analysis.}

\item{feature}{Graph feature that should be plotted}

\item{criterion}{Criterion proportion to compare the bootstrap proportion against.}
}
\description{
Plots bootstrap confidence intervals for each feature identified in the graph
and compares it against a predefined crtiterion.
}
\examples{

# plot_bootstrap_results(Edges, "Edges")

}
