% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci_precision.R
\name{plot_results}
\alias{plot_results}
\title{Helper function to plot simulation results, to be used within fci_precision}
\usage{
plot_results(results_output, plot_line = TRUE, layout = "complete_m")
}
\arguments{
\item{results_output}{List of data frames holding the formated simulation results, created by format_results}

\item{plot_line}{Logical whether loess regression line should be added to result plot.}

\item{layout}{Layout for the plots illustrating the results of the simulation. Options are: "complete_m", "cycle_m", "dag_m", "complete_s", "cycle_s", "dag_s", "edges_s".}
}
\description{
Helper function to plot simulation results, to be used within fci_precision
}
