% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_results.R
\name{summary_stat}
\alias{summary_stat}
\title{Helper function to extract summary statistic, to be used within bootstrap_results.}
\usage{
summary_stat(data_frame_list, stat)
}
\arguments{
\item{data_frame_list}{Feature column of the bootrstrap_samples object created by fci_bootstrap}

\item{stat}{Statistics to be extracted, i.e. mean, sd}
}
\value{
Results data frame
}
\description{
Helper function to extract summary statistic, to be used within bootstrap_results.
}
\examples{
# summary_stat(bootstrap_samples[, feature], sd)
}
