% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undirectied_bootstrap_graph.R
\name{undirected_bootstrap_graph}
\alias{undirected_bootstrap_graph}
\title{Undirected Bootstrap Graph}
\usage{
undirected_bootstrap_graph(bootstrap_edges, sig = FALSE, threshold = 0)
}
\arguments{
\item{bootstrap_edges}{Dataframe holding the proportion of time each edge has been included created by bootstrap_results(_, "Edges")}

\item{sig}{Logical if true only edges are included that are identified significantly more often than 0 times.}

\item{threshold}{Value between 0 and 1. If sig is false only edges are included with a proportion significantly greater than the threshold.}
}
\description{
Plots undirected bootsrap graph in which the proportion of times an eges has been
included corresponds to the saturation of the edge.
}
\examples{
# undirected_bootstrap_graph(Edges_results, threshold = .8)

}
